Remove workMu from tcpip.Endpoint.
workMu is removed and e.mu is now a tmutex that supports TryLock.
The packet processing path tries to lock the mutex and if its locked
it will just queue the packet and move on. The endpoint mutex now
on its Unlock() will process any backlog of packets before unlocking
the socket.

This simplifies the locking inside tcp endpoints a lot. Further the
endpoint.Lock() implements spinning as long as the lock is not held
by another syscall goroutine. This ensures low latency as not spinning
leads to the task thread being put to sleep if the lock is held
by the packet dispatch path. This is suboptimal as the lower layer
rarely holds the lock for long so implementing spinning here helps.

If the lock is held by another task goroutine then we just proceed
to call Lock() and the task could be put to sleep.

The protocol goroutines themselves just call e.mu.Lock() and block if
the lock is currently not available.

Updates #231

PiperOrigin-RevId: 285849425